create users.go
